requirements.txt
numpy
pandas
scipy
matplotlib
Pillow
scikit-learn
scikit-image
opencv-python
tensorflow
keras
